+++ src/main/java/coderank/impl/analyzer/AnalyzerNode.java
@@ -1,7 +1,15 @@
 package coderank.impl.analyzer;
 
-public interface AnalyzerNode {
-    double getRank();
-    void setRank(double rank);
-    int getIndex();
+public class AnalyzerNode {
+    public int index;
+    private double rank = 0;
+    public double getRank() {
+        return rank;
+    }
+    public void setRank(double rank) {
+        this.rank = rank;
+    }
+    public int getIndex() {
+        return index;
+    }
 }
