+++ src/main/java/coderank/impl/katzcentrality/KatzCentralityLauncher.java
@@ -1,13 +1,13 @@
 package coderank.impl.katzcentrality;
 
 import coderank.impl.analyzer.AnalyzerLauncher;
-import coderank.impl.analyzer.AnalyzerNode;
-import coderank.impl.javagraph.MethodNode;
 import coderank.impl.javagraph.Node;
 
-import java.io.*;
-import java.nio.charset.StandardCharsets;
-import java.util.*;
+import java.io.OutputStreamWriter;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+
 
 public class KatzCentralityLauncher<T> implements AnalyzerLauncher<T> {
     public void launch(HashSet<Node<T>> initStorage, HashMap<Node<T>, List<Node<T>>> edges,
