+++ src/main/java/coderank/impl/pagerank/PageGraph.java
@@ -3,10 +3,9 @@ package coderank.impl.pagerank;
 import coderank.impl.analyzer.Analyzer;
 import coderank.impl.javagraph.Node;
 
-import java.io.*;
-import java.nio.charset.StandardCharsets;
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
 
 public class PageGraph<T> implements Analyzer<T> {
     private static int index = 0;
